Extract asserts to mixin
